ring_buffer

Struct RingBuffer

pub struct RingBuffer<A: Copy, const CAP: usize> { /* private fields */ }

Very simple ring buffer that can hold up to CAP items of type A.

Implementations

impl<A: Copy, const CAP: usize> RingBuffer<A, CAP>

pub const fn new() -> Self

Creates a new empty ring buffer.

pub const fn len(&self) -> usize

Returns the number of items in the ring buffer.

pub const fn is_empty(&self) -> bool

Returns true if the ring buffer is empty.

pub const fn remaining(&self) -> usize

Returns the number of items that can be added to the ring buffer before it is full.

Same as CAP - self.len().

Examples

let mut buf = RingBuffer::<_, 3>::new();
assert_eq!(buf.remaining(), 3);
buf.extend([0, 1]);
assert_eq!(buf.remaining(), 1);
buf.push(2);
assert_eq!(buf.remaining(), 0);
assert_eq!(buf.pop_first(), Some(0));
assert_eq!(buf.remaining(), 1);

pub const fn has_remaining(&self) -> bool

Returns true if the ring buffer has capacity for at least one more item.

Equivalent to self.remaining() > 0.

pub unsafe fn get_unchecked(&self, index: usize) -> &A

Returns a reference to the item at the given index without doing bounds checks. Also assumes that the item is initialized.

Safety

The given index must be less than self.len().

pub unsafe fn get_unchecked_mut(&mut self, index: usize) -> &mut A

Returns a mutable reference to the item at the given index without doing bounds checks. Also assumes that the item is initialized.

Safety

The given index must be less than self.len().

pub fn get(&self, index: usize) -> Option<&A>

Returns a reference to the item at the given index, or None if the index is out of bounds.

Examples

let buf = RingBuffer::<_, 3>::from([0, 1]);
assert_eq!(buf.get(0), Some(&0));
assert_eq!(buf.get(1), Some(&1));
assert_eq!(buf.get(2), None);

pub fn get_mut(&mut self, index: usize) -> Option<&mut A>

Returns a mutable reference to the item at the given index, or None if the index is out of bounds.

Examples

let mut buf = RingBuffer::<_, 3>::from([1, 2]);
*buf.get_mut(0).unwrap() *= 2;
*buf.get_mut(1).unwrap() *= 3;
assert_eq!(buf.get_mut(2), None);
assert_eq!(buf, [2, 6]);

pub fn pop_first(&mut self) -> Option<A>

Removes the first item from the ring buffer and returns it, or None if the buffer is empty.

Examples

let mut buf = RingBuffer::<_, 3>::new();
assert_eq!(buf.pop_first(), None);
buf.extend([0, 1]);
assert_eq!(buf.pop_first(), Some(0));
assert_eq!(buf.pop_first(), Some(1));
assert_eq!(buf.pop_first(), None);

pub fn with_vacancy(&mut self) -> Option<VacantEntry<'_, A, CAP>>

If the ring has remaining capacity, returns a VacantEntry that can be used to push a new item to the end of the ring buffer.

See also try_push and try_push_with.

Examples

let mut buf = RingBuffer::<_, 3>::from([1, 2]);
buf.with_vacancy().unwrap().write(3);
assert_eq!(buf, [1, 2, 3]);
assert!(buf.with_vacancy().is_none());

pub fn pop_push(&mut self, item: A) -> Option<A>

Adds an item to the end of the ring buffer, removing the first item if the buffer has no remaining capacity. Returns the removed item if the buffer was full, otherwise None.

Panics

Panics if CAP is 0.

Examples

let mut buf = RingBuffer::<_, 3>::new();
assert_eq!(buf.pop_push(0), None);
assert_eq!(buf, [0]);
assert_eq!(buf.pop_push(1), None);
assert_eq!(buf, [0, 1]);
assert_eq!(buf.pop_push(2), None);
assert_eq!(buf, [0, 1, 2]);
assert_eq!(buf.pop_push(3), Some(0));
assert_eq!(buf, [1, 2, 3]);
assert_eq!(buf.pop_push(4), Some(1));
assert_eq!(buf, [2, 3, 4]);

pub fn try_push_with<F: FnOnce() -> A>( &mut self, f: F, ) -> Result<(), BufferFullError>

Tries to add an item to the end of the ring buffer, if the buffer has remaining capacity. Returns Ok(()) if successful, otherwise returns Err(BufferFullError).

The item is constructed by calling the given closure, which is only called if the buffer has remaining capacity.

Examples

let mut buf = RingBuffer::<_, 3>::from([0, 1]);
let mut call_count = 0;
assert!(
    buf.try_push_with(|| {
        call_count += 1;
        2
    }
).is_ok());
assert_eq!(call_count, 1);
assert_eq!(buf, [0, 1, 2]);
assert!(
    buf.try_push_with(|| {
        call_count += 1;
        3
    }
).is_err());
assert_eq!(call_count, 1);

pub unsafe fn push_unchecked(&mut self, item: A)

Adds an item to the end of the ring buffer, assuming the buffer has remaining capacity.

Safety

The following invariant must be held:

self.has_remaining() (which implies CAP > 0)

pub fn try_push(&mut self, item: A) -> Result<(), BufferFullError>

Tries to add an item to the end of the ring buffer, if the buffer has remaining capacity. Returns Ok(()) if successful, otherwise returns Err(BufferFullError).

Examples

let mut buf = RingBuffer::<_, 3>::from([1, 2]);
assert_eq!(buf.try_push(3), Ok(()));
assert_eq!(buf, [1, 2, 3]);
assert!(buf.try_push(4).is_err());

pub fn push(&mut self, item: A)

Adds an item to the end of the ring buffer.

Panics

Panics if the buffer has no remaining capacity.

Examples

let mut buf = RingBuffer::<_, 3>::new();
buf.push(0);
assert_eq!(buf, [0]);
buf.push(1);
assert_eq!(buf, [0, 1]);
buf.push(2);
assert_eq!(buf, [0, 1, 2]);

let mut buf = RingBuffer::<_, 3>::from([0, 1, 2]);
buf.push(3);

pub fn iter(&self) -> Iter<'_, A, CAP>

Returns an iterator over the items in the ring buffer.

pub fn iter_mut(&mut self) -> IterMut<'_, A, CAP>

Returns an iterator over the mutable references to the items in the ring buffer.

pub fn clear(&mut self)

Removes all items from the ring buffer.

pub unsafe fn as_slice_unchecked(&self) -> &[A]

Returns a slice containing the contents of the ring buffer, assuming ring buffer is layed out contiguously in memory.

Safety

The following invariant must be held:

self.pos.is_contiguous()

pub unsafe fn as_slice_unchecked_mut(&mut self) -> &mut [A]

Returns a mutable slice containing the contents of the ring buffer, assuming ring buffer is layed out contiguously in memory.

Safety

The following invariant must be held:

self.pos.is_contiguous()

pub fn as_slice(&self) -> Option<&[A]>

Returns a slice reference containing the contents of the ring buffer, if the ring buffer is layed out contiguously in memory.

Examples

let mut buf = RingBuffer::<_, 3>::from([0, 1]);
assert_eq!(buf.as_slice(), Some(&[0, 1][..]));
buf.push(2);
assert_eq!(buf.as_slice(), Some(&[0, 1, 2][..]));
buf.pop_first();
assert_eq!(buf.as_slice(), Some(&[1, 2][..]));
buf.push(3);
assert_eq!(buf.as_slice(), None);
buf.extend([4, 5]);
assert_eq!(buf.as_slice(), Some(&[3, 4, 5][..]));

pub fn as_slice_mut(&mut self) -> Option<&mut [A]>

Returns a slice mutable slice reference containing the contents of the ring buffer, if the ring buffer is layed out contiguously in memory.

Examples

let mut buf = RingBuffer::<_, 3>::from([0, 1]);
buf.as_slice_mut().unwrap().copy_from_slice(&[2, 3]);
assert_eq!(buf, [2, 3]);
buf.extend([4, 5]);
assert_eq!(buf.as_slice_mut(), None);
buf.extend([6, 7]);
buf.as_slice_mut().unwrap().copy_from_slice(&[8, 9, 10]);
assert_eq!(buf, [8, 9, 10]);

pub fn front_slice(&self) -> &[A]

Returns a slice reference containing the front part of the ring buffer.

If the ring buffer is not layed out contiguously in memory, this slice will not contain all the items, and the rest of the items will be in the back part of the ring buffer.

Examples

let mut buf = RingBuffer::<_, 3>::from([0, 1, 2]);
assert_eq!(buf.front_slice(), &[0, 1, 2][..]);
assert_eq!(buf.pop_first(), Some(0));
assert_eq!(buf.front_slice(), &[1, 2][..]);
buf.push(3);
assert_eq!(buf.front_slice(), &[1, 2][..]);
buf.pop_push(4);
assert_eq!(buf.front_slice(), &[2][..]);

pub fn front_slice_mut(&mut self) -> &mut [A]

Returns a mutable slice reference containing the front part of the ring buffer.

If the ring buffer is not layed out contiguously in memory, this slice will not contain all the items, and the rest of the items will be in the back part of the ring buffer.

See also split_mut_slices.

pub fn back_slice(&self) -> &[A]

Returns a slice reference containing the back part of the ring buffer.

If the ring buffer is layed out contiguously in memory, this slice will be empty. Otherwise, it will contain all the items that are not in the front part of the ring buffer.

Examples

let mut buf = RingBuffer::<_, 3>::from([0, 1, 2]);
assert_eq!(buf.back_slice(), &[][..]);
// at this point, the next item will be written to the back part of the buffer
assert_eq!(buf.pop_push(3), Some(0));
assert_eq!(buf.back_slice(), &[3][..]);

pub fn back_slice_mut(&mut self) -> &mut [A]

Returns a mutable slice reference containing the back part of the ring buffer.

If the ring buffer is layed out contiguously in memory, this slice will be empty. Otherwise, it will contain all the items that are not in the front part of the ring buffer.

See also split_mut_slices.

pub fn split_mut_slices(&mut self) -> (&mut [A], &mut [A])

Splits the ring buffer into two mutable slice references, one containing the front part of the ring buffer, and one containing the back part.

If the ring buffer is layed out contiguously in memory, the back slice will be empty, and the front slice will contain all the items.

Examples

let mut buf = RingBuffer::<_, 3>::new();
let (front, back) = buf.split_mut_slices();
assert_eq!(front, &[][..]);
assert_eq!(back, &[][..]);
buf.extend([0, 1, 2]);
let (front, back) = buf.split_mut_slices();
assert_eq!(front, &[0, 1, 2][..]);
assert_eq!(back, &[][..]);
buf.pop_push(3);
let (front, back) = buf.split_mut_slices();
front.copy_from_slice(&[4, 5]);
back.copy_from_slice(&[6]);
assert_eq!(buf, [4, 5, 6]);
buf.pop_push(7);
let (front, back) = buf.split_mut_slices();
front.copy_from_slice(&[8]);
back.copy_from_slice(&[9, 10]);
assert_eq!(buf, [8, 9, 10]);

pub fn to_array(&self) -> Option<[A; CAP]>

Copies the contents of the ring buffer into an array, if the ring buffer’s capacity is fully used. Otherwise, returns None.

Examples

assert_eq!(RingBuffer::<_, 3>::from([0, 1, 2]).to_array(), Some([0, 1, 2]));
assert_eq!(RingBuffer::<_, 3>::from([0, 1]).to_array(), None);

pub fn to_vec_in<B: Allocator>(&self, alloc: B) -> Vec<A>

Copies the contents of the ring buffer into a Vec, using the given allocator.

pub fn to_vec(&self) -> Vec<A>

Copies the contents of the ring buffer into a Vec.